refactor(types): move type and schemas to upper level folder
